move MSpaceMessageBox to bapsf_motion.gui.configure.message_boxes
